+++ pybamm/models/submodels/interface/base_interface.py
@@ -422,25 +422,6 @@ class BaseInterface(pybamm.BaseSubModel):
 
         return variables
 
-    def _get_standard_surface_potential_difference_variables(self, delta_phi):
-        domain, Domain = self.domain_Domain
-        ocp_ref = self.domain_param.U_ref
-
-        # Broadcast if necessary
-        delta_phi_dim = ocp_ref + delta_phi * self.param.potential_scale
-        if delta_phi.domain == ["current collector"]:
-            delta_phi = pybamm.PrimaryBroadcast(delta_phi, f"{domain} electrode")
-            delta_phi_dim = pybamm.PrimaryBroadcast(
-                delta_phi_dim, f"{domain} electrode"
-            )
-
-        variables = {
-            f"{Domain} electrode surface potential difference": delta_phi,
-            f"{Domain} electrode surface potential difference [V]": delta_phi_dim,
-        }
-
-        return variables
-
     def _get_standard_size_distribution_interfacial_current_variables(self, j):
         """
         Interfacial current density variables that depend on particle size R,